fix: widen classIndex/classObjClassIdx/excludePairs from short[] to int[]
short[] can only hold classList indices 0..32767, silently dropping all
classes with index ≥ 32768. The customer dump (pc52bs2job, 133K classes)
exposed this: java.util.HashMap and tens of thousands of other common
classes had classIndex=-1 (overflow), making them invisible in the
class histogram.

Changes:
- HeapGraph: short[] classIndex → int[], short[] classObjClassIdx → int[],
  short[][] excludePairs → int[][]
- HeapGraphBuilder: flushClassIndex returns int[], buildClassObjClassIdx
  removes Short.MAX_VALUE clamp, NamedEdgeConsumer takes int srcClassIdx,
  isExcluded/resolveExcludePairs updated to int[]
- All consumers (CsrBuilder, LeakSuspectsReport, TopConsumersReport,
  SystemOverviewReport, RetainedSizes, HtmlReportData): short ci → int ci
- HtmlReportData.loaderName: add guard for classIdx < 0 (class objects)
- DominatorTree: gate early inbound-CSR free on !retainInboundCsrForTesting
- HeapGraphBuilder.buildInternal: retainInboundCsr param plumbed to
  buildForTesting() so unit tests can still inspect inboundOffsets after build

+++ src/main/java/me/bechberger/hprof/views/CsrBuilder.java
@@ -98,7 +98,7 @@ final class CsrBuilder {
      * @param nameIdx    interned field name index (0 = no name / array element)
      * @param srcClassIdx class index of src (for exclude pair matching)
      */
-    void addEdge(int src, int dst, short nameIdx, short srcClassIdx) {
+    void addEdge(int src, int dst, short nameIdx, int srcClassIdx) {
         int[] offsets = offsetsCursor;
         int pos = offsets[dst];
         if (isInbound && lowMemory && isExcluded(srcClassIdx, nameIdx)) {
@@ -115,12 +115,12 @@ final class CsrBuilder {
         fwdTargets[fwdOffsets[src]++] = dst;
     }
 
-    private boolean isExcluded(short classIdx, short nameIdx) {
+    private boolean isExcluded(int classIdx, short nameIdx) {
         // Class meta edges (instance→classObj, classObj→superClass/classLoader) are pseudo
         // references that MAT skips during dominator tree computation.
         if (nameIdx == Short.MIN_VALUE) return true;
         if (graph.excludePairs == null || nameIdx == ClassRecord.NO_NAME) return false;
-        for (short[] pair : graph.excludePairs) {
+        for (int[] pair : graph.excludePairs) {
             if (pair[0] == classIdx && pair[1] == nameIdx) return true;
         }
         return false;
